// frontend-v3/src/app/core/api/models/categoria-opcao-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Dados de saida de uma opcao de categoria
 */
export interface CategoriaOpcaoOutput {

  /**
   * Indica se a opcao esta ativa (soft delete)
   */
  ativo: boolean;

  /**
   * ID da categoria a que pertence a opcao
   */
  id_categoria: number;

  /**
   * ID da opcao de categoria
   */
  id_opcao: number;

  /**
   * Nome da opcao
   */
  nome: string;
}