// frontend-v3/src/app/core/api/models/produto-input.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ProdutoOpcaoInput } from '../models/produto-opcao-input';

/**
 * Dados de entrada do produto
 */
export interface ProdutoInput {

  /**
   * Descricao do produto
   */
  descricao?: string;

  /**
   * ID da categoria do produto
   */
  id_categoria: number;

  /**
   * Nome do produto
   */
  nome: string;

  /**
   * Lista de opcoes de preco para o produto
   */
  opcoes: Array<ProdutoOpcaoInput>;
}